Text Box addition
Added a textbox that will receive the abstract
