+++ webApp/prediction_ui/src/componenets/textBox.js
@@ -1,0 +1,17 @@
+// src/textBox.js
+import React from 'react';
+
+function textBox() {
+    return (
+        <div>
+        <input
+            type="text"
+            value={this.state.value}
+            onChange={() => console.log(this.state.value)}
+         />
+         </div>
+     );
+
+}
+
+export default textBox
